refactor(matcha-window): split adapter platform methods into native/web
Moves `Adapter::run` (return type diverges between native and wasm) and
`Adapter::with_tokio_runtime*` (native-only) into
`adapter/native.rs` and `adapter/web.rs`, mounted via `#[path]`.
`adapter.rs` keeps the struct definition and all common methods and is
now cfg-free.
